requirements.txt
Django==1.8.2
argparse==1.2.1
djangorestframework==3.1.3
djangorestframework-gis==0.9.2
factory-boy==2.5.2
mccabe==0.3.1
psycopg2==2.6.1
requests==2.7.0
stevedore==0.15
wsgiref==0.1.2
